modification of the back and initialization of the front

+++ src/main/java/fr/najet/bank/controllers/AuthController.java
@@ -57,7 +57,7 @@ public class AuthController {
         .collect(Collectors.toList());
 
     return ResponseEntity.ok(new JwtResponse(jwt,
-        userDetails.getId(),
+        (long) userDetails.getId(),
         userDetails.getUsername(),
         userDetails.getEmail(),
         roles));
